Got rid of assert in most places, fixed Makefile to work with bear, fixed small errors

+++ tests/test_crash_threaded_hashmap.c
@@ -22,6 +22,6 @@ int main()
     for (long long i = 0;;) {
         char key[32] = {};
         int len = snprintf(key, sizeof(key), "%lld", i++);
-        if (!m->ops->kevue_hm_put(m, key, len, ".", 1)) return 0;
+        if (!m->ops->kevue_hm_put(m, key, (size_t)len, ".", 1)) return 0;
     }
 }
